requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
aiomysql==0.2.0

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# 环境变量
python-dotenv==1.0.0

# 异步支持
greenlet==3.0.1

# 工具
python-multipart==0.0.6